Made some changes that Pornkamol requested
- Added an API to delete threads (not tested, may need to use another
complicated method in case deleting threads by id doesn't work)
- Added .filter(Comment.main_comment) to get_top_comment() since replies
could have more upvotes than main comments
- Added Boolean value 'deleted' to Comment model (initialized as True when comment is created and set to False when deleted). Also sent this value to frontend in query.py so they can check to see if the comment is deleted and put some italicized text saying "This comment has been deleted" or something like that
- Added checks to see if the user editing/deleting a thread/comment is the owner of said thread/comment, just to be safe

+++ src/routes/comments_api.py
@@ -31,25 +31,33 @@ def new_comment():
 @app.route('/api/edit_comment')
 def edit_comment():
 
-    new_comment_body = request.form.get("comment_body")
-
-    if (len(new_comment_body.strip()) == 0):
-        return jsonify(status=False, message="Empty comment body")
-
-    # Again, defaulting to request.form.get until we have a way to request
     comment_id = request.form.get("comment_id")
-    write_queries.edit_comment(comment_id, new_comment_body)
-    return jsonify(status=True, message="Reply created successfully")
+    username = request.form.get("sky_username")
+    comment = read_queries.get_comment_by_id(comment_id)
+    new_comment_body = request.form.get("comment_body")
+    
+    if (username == comment.user_id):
+        if (len(new_comment_body.strip()) == 0):
+            return jsonify(status=False, message="Empty comment body")
 
+        write_queries.edit_comment(comment_id, new_comment_body)
+        return jsonify(status=True, message="Reply created successfully")
+    
+    return jsonify(status=False, message="Unable to edit: user is not the owner of the comment")
 
 @app.route("/api/delete_comment")
 def delete_comment():
 
     # Again, defaulting to request.form.get until we have a way to request
     comment_id = request.form.get("comment_id")
-    write_queries.delete_comment(comment_id)
-    return jsonify(status=True, message="Comment deleted successfully")
+    username = request.form.get("sky_username")
+    comment = read_queries.get_comment_by_id(comment_id)
 
+    if (username == comment.user_id):
+        write_queries.delete_comment(comment_id)
+        return jsonify(status=True, message="Comment deleted successfully")
+    
+    return jsonify(status=False, message="Unable to delete: user is not the owner of the comment")
 
 @app.route("/api/like_comment", methods=["POST"])
 def like_comment():
